Enable -Wconversion.

+++ src/cbpp/status_codes.hpp
@@ -7,7 +7,7 @@
 namespace Poseidon {
 namespace Cbpp {
 
-typedef int StatusCode;
+typedef long StatusCode;
 
 namespace StatusCodes {
 	enum {
